+++ pkg/externalrepo/git/git.go
@@ -298,469 +298,6 @@ func (r *gitRepository) DeletePackage(ctx context.Context, obj repository.Packag
 	return fmt.Errorf("DeletePackage not yet supported for git repos")
 }
 
-func (r *gitRepository) ListPackageRevisions(ctx context.Context, filter repository.ListPackageRevisionFilter) ([]repository.PackageRevision, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::ListPackageRevisions", trace.WithAttributes())
-	defer span.End()
-
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	pkgRevs, err := r.listPackageRevisions(ctx, filter)
-	if err != nil {
-		return nil, err
-	}
-	var repoPkgRevs []repository.PackageRevision
-	for i := range pkgRevs {
-		repoPkgRevs = append(repoPkgRevs, pkgRevs[i])
-	}
-	return repoPkgRevs, nil
-}
-
-func (r *gitRepository) listPackageRevisions(ctx context.Context, filter repository.ListPackageRevisionFilter) ([]*gitPackageRevision, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::listPackageRevisions", trace.WithAttributes())
-	defer span.End()
-
-	if err := r.fetchRemoteRepository(ctx); err != nil {
-		return nil, err
-	}
-
-	refs, err := r.repo.References()
-	if err != nil {
-		return nil, err
-	}
-
-	var main *plumbing.Reference
-	var drafts []*gitPackageRevision
-	var result []*gitPackageRevision
-
-	mainBranch := r.branch.RefInLocal() // Looking for the registered branch
-
-	for ref, err := refs.Next(); err == nil; ref, err = refs.Next() {
-		switch name := ref.Name(); {
-		case name == mainBranch:
-			main = ref
-			continue
-
-		case isProposedBranchNameInLocal(ref.Name()), isDraftBranchNameInLocal(ref.Name()):
-			klog.Infof("Loading draft from %q", ref.Name())
-			draft, err := r.loadDraft(ctx, ref)
-			if err != nil {
-				return nil, pkgerrors.Wrapf(err, "failed to load package draft %q", name.String())
-			}
-			if draft != nil {
-				drafts = append(drafts, draft)
-			} else {
-				klog.Warningf("no package draft found for ref %v", ref)
-			}
-		case isTagInLocalRepo(ref.Name()):
-			klog.Infof("Loading tag from %q", ref.Name())
-			tagged, err := r.loadTaggedPackage(ctx, ref)
-			if err != nil {
-				if tagged == nil {
-					// this tag is not associated with any package (e.g. could be a release tag)
-					klog.Warningf("Failed to load tagged package from ref %q: %s", ref.Name(), err)
-					continue
-				}
-				klog.Warningf("Error loading tagged package from ref %q: %s", ref.Name(), err)
-			}
-			if tagged != nil && filter.Matches(ctx, tagged) {
-				result = append(result, tagged)
-			}
-		}
-	}
-
-	if main != nil {
-		// TODO: ignore packages that are unchanged in main branch, compared to a tagged version?
-		mainpkgs, err := r.discoverFinalizedPackages(ctx, main)
-		if err != nil {
-			if len(mainpkgs) == 0 {
-				return nil, err
-			}
-			klog.Warningf("Error discovering finalized packages: %s", err)
-		}
-		for _, p := range mainpkgs {
-			if filter.Matches(ctx, p) {
-				result = append(result, p)
-			}
-		}
-	}
-
-	for _, p := range drafts {
-		if filter.Matches(ctx, p) {
-			result = append(result, p)
-		}
-	}
-
-	return result, nil
-}
-
-func (r *gitRepository) CreatePackageRevision(ctx context.Context, obj *v1alpha1.PackageRevision) (repository.PackageRevisionDraft, error) {
-	_, span := tracer.Start(ctx, "gitRepository::CreatePackageRevision", trace.WithAttributes())
-	defer span.End()
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	var base plumbing.Hash
-	refName := r.branch.RefInLocal()
-	switch main, err := r.repo.Reference(refName, true); {
-	case err == nil:
-		base = main.Hash()
-	case err == plumbing.ErrReferenceNotFound:
-		// reference not found - empty repository. Package draft has no parent commit
-	default:
-		return nil, fmt.Errorf("error when resolving target branch for the package: %w", err)
-	}
-
-	if err := util.ValidPkgRevObjName(r.Key().Name, r.Key().Path, obj.Spec.PackageName, string(obj.Spec.WorkspaceName)); err != nil {
-		return nil, fmt.Errorf("failed to create packagerevision: %w", err)
-	}
-
-	draftKey := repository.PackageRevisionKey{
-		PkgKey:        repository.FromFullPathname(r.Key(), obj.Spec.PackageName),
-		WorkspaceName: obj.Spec.WorkspaceName,
-	}
-
-	// TODO use git branches to leverage uniqueness
-	draftBranchName := createDraftName(draftKey)
-
-	// TODO: This should also create a new 'Package' resource if one does not already exist
-
-	return &gitPackageRevisionDraft{
-		prKey:     draftKey,
-		repo:      r,
-		lifecycle: v1alpha1.PackageRevisionLifecycleDraft,
-		updated:   time.Now(),
-		base:      nil, // Creating a new package
-		tasks:     nil, // Creating a new package
-		branch:    draftBranchName,
-		commit:    base,
-	}, nil
-}
-
-func (r *gitRepository) UpdatePackageRevision(ctx context.Context, old repository.PackageRevision) (repository.PackageRevisionDraft, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::UpdatePackageRevision", trace.WithAttributes())
-	defer span.End()
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	oldGitPackage, ok := old.(*gitPackageRevision)
-	if !ok {
-		return nil, fmt.Errorf("cannot update non-git package %T", old)
-	}
-
-	ref := oldGitPackage.ref
-	if ref == nil {
-		return nil, fmt.Errorf("cannot update final package")
-	}
-
-	head, err := r.repo.Reference(ref.Name(), true)
-	if err != nil {
-		return nil, fmt.Errorf("cannot find draft package branch %q: %w", ref.Name(), err)
-	}
-
-	rev, err := r.loadDraft(ctx, head)
-	if err != nil {
-		return nil, fmt.Errorf("cannot load draft package: %w", err)
-	}
-	if rev == nil {
-		return nil, fmt.Errorf("cannot load draft package %q (package not found)", ref.Name())
-	}
-
-	// Fetch lifecycle directly from the repository rather than from the gitPackageRevision. This makes
-	// sure we don't end up requesting the same lock twice.
-	lifecycle := r.getLifecycle(oldGitPackage)
-
-	return &gitPackageRevisionDraft{
-		prKey:     oldGitPackage.prKey,
-		repo:      r,
-		lifecycle: lifecycle,
-		updated:   rev.updated,
-		base:      rev.ref,
-		tree:      rev.tree,
-		commit:    rev.commit,
-		tasks:     rev.tasks,
-	}, nil
-}
-
-func (r *gitRepository) DeletePackageRevision(ctx context.Context, old repository.PackageRevision) error {
-	ctx, span := tracer.Start(ctx, "gitRepository::DeletePackageRevision", trace.WithAttributes())
-	defer span.End()
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	oldGit, ok := old.(*gitPackageRevision)
-	if !ok {
-		return fmt.Errorf("cannot delete non-git package: %T", old)
-	}
-
-	ref := oldGit.ref
-	if ref == nil {
-		// This is an internal error. In some rare cases (see GetPackageRevision below) we create
-		// package revisions without refs. They should never be returned via the API though.
-		return fmt.Errorf("cannot delete package with no ref: %s", oldGit.Key().PkgKey.ToFullPathname())
-	}
-
-	// We can only delete packages which have their own ref. Refs that are shared with other packages
-	// (main branch, tag that doesn't contain package path in its name, ...) cannot be deleted.
-
-	refSpecs := newPushRefSpecBuilder()
-
-	switch rn := ref.Name(); {
-	case rn.IsTag():
-		// Delete tag only if it is package-specific.
-		name := createFinalTagNameInLocal(oldGit.Key())
-		if rn != name {
-			return fmt.Errorf("cannot delete package tagged with a tag that is not specific to the package: %s", rn)
-		}
-
-		// Delete the tag
-		refSpecs.AddRefToDelete(ref)
-
-		// If this revision was proposed for deletion, we need to delete the associated branch.
-		if err := r.removeDeletionProposedBranchIfExists(ctx, oldGit.Key()); err != nil {
-			return err
-		}
-
-	case isDraftBranchNameInLocal(rn), isProposedBranchNameInLocal(rn):
-		// PackageRevision is proposed or draft; delete the branch directly.
-		refSpecs.AddRefToDelete(ref)
-
-	case isBranchInLocalRepo(rn):
-		// Delete package from the branch
-		commitHash, err := r.createPackageDeleteCommit(ctx, rn, oldGit)
-		if err != nil {
-			return err
-		}
-
-		// Remove the proposed for deletion branch. We end up here when users
-		// try to delete the main branch version of a packagerevision.
-		if err := r.removeDeletionProposedBranchIfExists(ctx, oldGit.Key()); err != nil {
-			return err
-		}
-
-		// Update the reference
-		refSpecs.AddRefToPush(commitHash, rn)
-
-	default:
-		return fmt.Errorf("cannot delete package with the ref name %s", rn)
-	}
-
-	// Update references
-	if err := r.pushAndCleanup(ctx, refSpecs); err != nil {
-		return fmt.Errorf("failed to update git references: %v", err)
-	}
-
-	return nil
-}
-
-func (r *gitRepository) removeDeletionProposedBranchIfExists(ctx context.Context, key repository.PackageRevisionKey) error {
-	refSpecsForDeletionProposed := newPushRefSpecBuilder()
-	deletionProposedBranch := createDeletionProposedName(key)
-	refSpecsForDeletionProposed.AddRefToDelete(plumbing.NewHashReference(deletionProposedBranch.RefInLocal(), plumbing.ZeroHash))
-	if err := r.pushAndCleanup(ctx, refSpecsForDeletionProposed); err != nil {
-		if pkgerrors.Is(err, git.NoErrAlreadyUpToDate) {
-			// the deletionProposed branch might not have existed, so we ignore this error
-			klog.Warningf("branch %s does not exist", deletionProposedBranch)
-		} else {
-			klog.Errorf("unexpected error while removing deletionProposed branch: %v", err)
-			return err
-		}
-	}
-	delete(r.deletionProposedCache, deletionProposedBranch)
-
-	return nil
-}
-
-func (r *gitRepository) GetPackageRevision(ctx context.Context, version, path string) (repository.PackageRevision, kptfilev1.GitLock, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::GetPackageRevision", trace.WithAttributes())
-	defer span.End()
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	var hash plumbing.Hash
-
-	// Trim leading and trailing slashes
-	path = strings.Trim(path, "/")
-
-	// Versions map to gitRepo tags in one of two ways:
-	//
-	// * directly (tag=version)- but then this means that all packages in the repo must be versioned together.
-	// * prefixed (tag=<packageDir/<version>) - solving the co-versioning problem.
-	//
-	// We have to check both forms when looking up a version.
-	refNames := []string{}
-	if path != "" {
-		refNames = append(refNames, path+"/"+version)
-		// HACK: Is this always refs/remotes/origin ?  Is it ever not (i.e. do we need both forms?)
-		refNames = append(refNames, "refs/remotes/origin/"+path+"/"+version)
-	}
-	refNames = append(refNames, version)
-	// HACK: Is this always refs/remotes/origin ?  Is it ever not (i.e. do we need both forms?)
-	refNames = append(refNames, "refs/remotes/origin/"+version)
-
-	for _, ref := range refNames {
-		if resolved, err := r.repo.ResolveRevision(plumbing.Revision(ref)); err != nil {
-			if pkgerrors.Is(err, plumbing.ErrReferenceNotFound) {
-				continue
-			}
-			return nil, kptfilev1.GitLock{}, pkgerrors.Wrapf(err, "error resolving git reference %q", ref)
-		} else {
-			hash = *resolved
-			break
-		}
-	}
-
-	if hash.IsZero() {
-		r.dumpAllRefs()
-
-		return nil, kptfilev1.GitLock{}, pkgerrors.Errorf("cannot find git reference (tried %v)", refNames)
-	}
-
-	return r.loadPackageRevision(ctx, version, path, hash)
-}
-
-func (r *gitRepository) loadPackageRevision(ctx context.Context, version, path string, hash plumbing.Hash) (repository.PackageRevision, kptfilev1.GitLock, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::loadPackageRevision", trace.WithAttributes())
-	defer span.End()
-
-	if !packageInDirectory(path, r.Key().Path) {
-		return nil, kptfilev1.GitLock{}, fmt.Errorf("cannot find package %s@%s; package is not under the Repository.spec.directory", path, version)
-	}
-
-	origin, err := r.repo.Remote("origin")
-	if err != nil {
-		return nil, kptfilev1.GitLock{}, pkgerrors.Wrap(err, "cannot determine repository origin")
-	}
-
-	lock := kptfilev1.GitLock{
-		Repo:      origin.Config().URLs[0],
-		Directory: path,
-		Ref:       version,
-	}
-
-	commit, err := r.repo.CommitObject(hash)
-	if err != nil {
-		return nil, lock, pkgerrors.Wrapf(err, "cannot resolve git reference %s (hash: %s) to commit", version, hash)
-	}
-	lock.Commit = commit.Hash.String()
-
-	krmPackage, err := r.findPackage(commit, path)
-	if err != nil {
-		return nil, lock, err
-	}
-
-	if krmPackage == nil {
-		return nil, lock, pkgerrors.Errorf("cannot find package %s@%s", path, version)
-	}
-
-	var ref *plumbing.Reference = nil // Cannot determine ref; this package will be considered final (immutable).
-
-	var revisionStr, workspace string
-	last := strings.LastIndex(version, "/")
-
-	if strings.HasPrefix(version, "drafts/") || strings.HasPrefix(version, "proposed/") {
-		// the passed in version is a ref to an unpublished package revision
-		workspace = version[last+1:]
-	} else {
-		// the passed in version is a ref to a published package revision
-		if version == string(r.branch) || last < 0 {
-			revisionStr = version
-		} else {
-			revisionStr = version[last+1:]
-		}
-		workspace = getPkgWorkspace(commit, krmPackage, ref)
-		if workspace == "" {
-			workspace = revisionStr
-		}
-	}
-
-	packageRevision, err := krmPackage.buildGitPackageRevision(ctx, revisionStr, workspace, ref)
-	if err != nil {
-		return nil, lock, err
-	}
-	return packageRevision, lock, nil
-}
-
-func (r *gitRepository) discoverFinalizedPackages(ctx context.Context, ref *plumbing.Reference) ([]*gitPackageRevision, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::discoverFinalizedPackages", trace.WithAttributes())
-	defer span.End()
-
-	commit, err := r.repo.CommitObject(ref.Hash())
-	if err != nil {
-		return nil, err
-	}
-
-	var revisionStr string
-	if rev, ok := getBranchNameInLocalRepo(ref.Name()); ok {
-		revisionStr = rev
-	} else if rev, ok = getTagNameInLocalRepo(ref.Name()); ok {
-		revisionStr = rev
-	} else {
-		// TODO: ignore the ref instead?
-		return nil, pkgerrors.Errorf("cannot determine revision from ref: %q", rev)
-	}
-
-	krmPackages, err := r.discoverPackagesInTree(commit, DiscoverPackagesOptions{FilterPrefix: r.Key().Path, Recurse: true})
-	if err != nil {
-		return nil, err
-	}
-
-	ec := errors.NewErrorCollector().WithSeparator(";").WithFormat("{%s}")
-	var result []*gitPackageRevision
-	for _, krmPackage := range krmPackages.packages {
-		workspace := getPkgWorkspace(commit, krmPackage, ref)
-		if workspace == "" {
-			klog.Warningf("Failed to get workspace name for package %q (will use revision name): %s", krmPackage.pkgKey, err)
-		}
-		packageRevision, err := krmPackage.buildGitPackageRevision(ctx, revisionStr, workspace, ref)
-		if err != nil {
-			ec.Add(pkgerrors.Wrapf(err, "failed to build git package revision for package %s", krmPackage.pkgKey))
-			continue
-		}
-		result = append(result, packageRevision)
-	}
-
-	return result, ec.Join()
-}
-
-// loadDraft will load the draft package.  If the package isn't found (we now require a Kptfile), it will return (nil, nil)
-func (r *gitRepository) loadDraft(ctx context.Context, ref *plumbing.Reference) (*gitPackageRevision, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::loadDraft", trace.WithAttributes())
-	defer span.End()
-
-	pkgPathAndName, workspaceName, err := parseDraftName(ref)
-	if err != nil {
-		return nil, err
-	}
-
-	// Only load drafts in the directory specified at repository registration.
-	if !packageInDirectory(pkgPathAndName, r.Key().Path) {
-		return nil, nil
-	}
-
-	commit, err := r.repo.CommitObject(ref.Hash())
-	if err != nil {
-		return nil, pkgerrors.Wrap(err, "cannot resolve draft branch to commit (corrupted repository?)")
-	}
-
-	krmPackage, err := r.findPackage(commit, pkgPathAndName)
-	if err != nil {
-		return nil, err
-	}
-
-	if krmPackage == nil {
-		klog.Warningf("draft package %q was not found", pkgPathAndName)
-		return nil, nil
-	}
-
-	packageRevision, err := krmPackage.buildGitPackageRevision(ctx, "0", workspaceName, ref)
-	if err != nil {
-		return nil, err
-	}
-
-	return packageRevision, nil
-}
-
 func (r *gitRepository) UpdateDeletionProposedCache() error {
 	r.mutex.Lock()
 	defer r.mutex.Unlock()
@@ -796,135 +333,6 @@ func (r *gitRepository) updateDeletionProposedCache() error {
 	}
 
 	return nil
-}
-
-func parseDraftName(draft *plumbing.Reference) (pkgPathAndName, workspaceName string, err error) {
-	refName := draft.Name()
-	var suffix string
-	if b, ok := getDraftBranchNameInLocal(refName); ok {
-		suffix = string(b)
-	} else if b, ok = getProposedBranchNameInLocal(refName); ok {
-		suffix = string(b)
-	} else {
-		return "", "", fmt.Errorf("invalid draft ref name: %q", refName)
-	}
-
-	revIndex := strings.LastIndex(suffix, "/")
-	if revIndex <= 0 {
-		return "", "", fmt.Errorf("invalid draft ref name; missing workspaceName suffix: %q", refName)
-	}
-	pkgPathAndName, workspaceName = suffix[:revIndex], suffix[revIndex+1:]
-	return pkgPathAndName, workspaceName, nil
-}
-
-func (r *gitRepository) loadTaggedPackage(ctx context.Context, tag *plumbing.Reference) (*gitPackageRevision, error) {
-	ctx, span := tracer.Start(ctx, "gitRepository::loadTaggedPackage", trace.WithAttributes())
-	defer span.End()
-
-	name, ok := getTagNameInLocalRepo(tag.Name())
-	if !ok {
-		return nil, pkgerrors.Errorf("invalid tag ref: %q", tag)
-	}
-	slash := strings.LastIndex(name, "/")
-
-	if slash < 0 {
-		// tag=<version>
-		// could be a release tag or something else, we ignore these types of tags
-		return nil, pkgerrors.Errorf("could not find slash in %q", name)
-	}
-
-	// tag=<package path>/version
-	path, revisionStr := name[:slash], name[slash+1:]
-
-	if !packageInDirectory(path, r.Key().Path) {
-		return nil, nil
-	}
-
-	resolvedHash, err := r.repo.ResolveRevision(plumbing.Revision(tag.Hash().String()))
-	if err != nil {
-		return nil, pkgerrors.Wrapf(err, "cannot resolve tag %q to git revision", name)
-	}
-	commit, err := r.repo.CommitObject(*resolvedHash)
-	if err != nil {
-		return nil, pkgerrors.Wrapf(err, "cannot resolve tag %q (hash: %q) to commit (corrupted repository?)", name, resolvedHash)
-	}
-
-	krmPackage, err := r.findPackage(commit, path)
-	if err != nil {
-		return nil, pkgerrors.Wrapf(err, "skipping %q; cannot find %q (corrupted repository?)", name, path)
-	}
-
-	if krmPackage == nil {
-		return nil, pkgerrors.Errorf("skipping %q: Kptfile not found", name)
-	}
-
-	workspaceName := getPkgWorkspace(commit, krmPackage, tag)
-	if workspaceName == "" {
-		klog.Warningf("Failed to get package workspace name for package %q (will use revision name)", name)
-	}
-
-	packageRevision, err := krmPackage.buildGitPackageRevision(ctx, revisionStr, workspaceName, tag)
-	if err != nil {
-		if packageRevision == nil {
-			return nil, err
-		}
-		klog.Warningf("Error building git package revision %q: %s", path, err)
-	}
-
-	return packageRevision, nil
-}
-
-func (r *gitRepository) dumpAllRefs() {
-	refs, err := r.repo.References()
-	if err != nil {
-		klog.Warningf("failed to get references: %v", err)
-	} else {
-		for {
-			ref, err := refs.Next()
-			if err != nil {
-				if err != io.EOF {
-					klog.Warningf("failed to get next reference: %v", err)
-				}
-				break
-			}
-			klog.Infof("ref %#v", ref.Name())
-		}
-	}
-
-	branches, err := r.repo.Branches()
-	if err != nil {
-		klog.Warningf("failed to get branches: %v", err)
-	} else {
-		for {
-			branch, err := branches.Next()
-			if err != nil {
-				if err != io.EOF {
-					klog.Warningf("failed to get next branch: %v", err)
-				}
-				break
-			}
-			klog.Infof("branch %#v", branch.Name())
-		}
-	}
-}
-
-// getAuthMethod fetches the credentials for authenticating to git. It caches the
-// credentials between calls and refresh credentials when the tokens have expired.
-func (r *gitRepository) getAuthMethod(ctx context.Context, forceRefresh bool) (transport.AuthMethod, error) {
-	// If no secret is provided, we try without any auth.
-	if r.secret == "" {
-		return nil, nil
-	}
-
-	if r.credential == nil || !r.credential.Valid() || forceRefresh {
-		if cred, err := r.credentialResolver.ResolveCredential(ctx, r.Key().Namespace, r.secret); err != nil {
-			return nil, fmt.Errorf("failed to obtain credential from secret %s/%s: %w", r.Key().Namespace, r.secret, err)
-		} else {
-			r.credential = cred
-		}
-	}
-
-	return r.credential.ToAuthMethod(), nil
 }
 
 func (r *gitRepository) GetRepo() (string, error) {
@@ -1044,94 +452,12 @@ func (r *gitRepository) createBranch(ctx context.Context, branch BranchName) err
 	return r.pushAndCleanup(ctx, refSpecs)
 }
 
-func (r *gitRepository) getCommit(h plumbing.Hash) (*object.Commit, error) {
-	return object.GetCommit(r.repo.Storer, h)
-}
-
 func (r *gitRepository) storeCommit(commit *object.Commit) (plumbing.Hash, error) {
 	eo := r.repo.Storer.NewEncodedObject()
 	if err := commit.Encode(eo); err != nil {
 		return plumbing.Hash{}, err
 	}
 	return r.repo.Storer.SetEncodedObject(eo)
-}
-
-// Creates a commit which deletes the package from the branch, and returns its commit hash.
-// If the branch doesn't exist, will return zero hash and no error.
-func (r *gitRepository) createPackageDeleteCommit(ctx context.Context, branch plumbing.ReferenceName, pkg *gitPackageRevision) (plumbing.Hash, error) {
-	var zero plumbing.Hash
-
-	local, err := refInRemoteFromRefInLocal(branch)
-	if err != nil {
-		return zero, err
-	}
-	// Fetch the branch
-	// TODO: Fetch only as part of conflict resolution & Retry
-	switch err := r.doGitWithAuth(ctx, func(auth transport.AuthMethod) error {
-		return r.repo.Fetch(&git.FetchOptions{
-			RemoteName: OriginName,
-			RefSpecs:   []config.RefSpec{config.RefSpec(fmt.Sprintf("+%s:%s", local, branch))},
-			Auth:       auth,
-			Tags:       git.NoTags,
-			CABundle:   r.caBundle,
-		})
-	}); err {
-	case nil, git.NoErrAlreadyUpToDate:
-		// ok
-	default:
-		return zero, fmt.Errorf("failed to fetch remote repository: %w", err)
-	}
-
-	// find the branch
-	ref, err := r.repo.Reference(branch, true)
-	if err != nil {
-		// branch doesn't exist, and therefore package doesn't exist either.
-		klog.Infof("Branch %q no longer exist, deleting a package from it is unnecessary", branch)
-		return zero, nil
-	}
-	commit, err := r.repo.CommitObject(ref.Hash())
-	if err != nil {
-		return zero, fmt.Errorf("failed to resolve main branch to commit: %w", err)
-	}
-	root, err := commit.Tree()
-	if err != nil {
-		return zero, fmt.Errorf("failed to find commit tree for %s: %w", ref, err)
-	}
-
-	// Find the package in the tree
-	switch _, err := root.FindEntry(pkg.Key().PkgKey.ToFullPathname()); err {
-	case object.ErrEntryNotFound:
-		// Package doesn't exist; no need to delete it
-		return zero, nil
-	case nil:
-		// found
-	default:
-		return zero, fmt.Errorf("failed to find package %q in the repositrory ref %q: %w,", pkg.Key().PkgKey.ToFullPathname(), ref, err)
-	}
-
-	// Create commit helper. Use zero hash for the initial package tree. Commit helper will initialize trees
-	// without TreeEntry for this package present - the package is deleted.
-	ch, err := newCommitHelper(r, r.userInfoProvider, commit.Hash, pkg.Key().PkgKey.ToFullPathname(), zero)
-	if err != nil {
-		return zero, fmt.Errorf("failed to initialize commit of package %q to %q: %w", pkg.Key().PkgKey.ToFullPathname(), ref, err)
-	}
-
-	message := fmt.Sprintf("Delete %s", pkg.Key().PkgKey.ToFullPathname())
-	commitHash, _, err := ch.commit(ctx, message, pkg.Key().PkgKey.ToFullPathname())
-	if err != nil {
-		return zero, fmt.Errorf("failed to commit package %q to %q: %w", pkg.Key().PkgKey.ToFullPathname(), ref, err)
-	}
-	return commitHash, nil
-}
-
-func (r *gitRepository) PushAndCleanup(ctx context.Context, ph *pushRefSpecBuilder) error {
-	ctx, span := tracer.Start(ctx, "gitRepository::PushAndCleanup", trace.WithAttributes())
-	defer span.End()
-
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	return r.pushAndCleanup(ctx, ph)
 }
 
 func (r *gitRepository) pushAndCleanup(ctx context.Context, ph *pushRefSpecBuilder) error {
@@ -1364,141 +690,6 @@ func (r *gitRepository) storeTree(tree *object.Tree) (plumbing.Hash, error) {
 	return treeHash, nil
 }
 
-func (r *gitRepository) GetLifecycle(ctx context.Context, pkgRev *gitPackageRevision) v1alpha1.PackageRevisionLifecycle {
-	_, span := tracer.Start(ctx, "gitRepository::GetLifecycle", trace.WithAttributes())
-	defer span.End()
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	return r.getLifecycle(pkgRev)
-}
-
-func (r *gitRepository) getLifecycle(pkgRev *gitPackageRevision) v1alpha1.PackageRevisionLifecycle {
-	switch ref := pkgRev.ref; {
-	case ref == nil:
-		return r.checkPublishedLifecycle(pkgRev)
-	case isDraftBranchNameInLocal(ref.Name()):
-		return v1alpha1.PackageRevisionLifecycleDraft
-	case isProposedBranchNameInLocal(ref.Name()):
-		return v1alpha1.PackageRevisionLifecycleProposed
-	default:
-		return r.checkPublishedLifecycle(pkgRev)
-	}
-}
-
-func (r *gitRepository) checkPublishedLifecycle(pkgRev *gitPackageRevision) v1alpha1.PackageRevisionLifecycle {
-	if r.deletionProposedCache == nil {
-		if err := r.updateDeletionProposedCache(); err != nil {
-			klog.Errorf("failed to update deletionProposed cache: %v", err)
-			return v1alpha1.PackageRevisionLifecyclePublished
-		}
-	}
-
-	branchName := createDeletionProposedName(pkgRev.Key())
-	if _, found := r.deletionProposedCache[branchName]; found {
-		return v1alpha1.PackageRevisionLifecycleDeletionProposed
-	}
-
-	return v1alpha1.PackageRevisionLifecyclePublished
-}
-
-func (r *gitRepository) UpdateLifecycle(ctx context.Context, pkgRev *gitPackageRevision, newLifecycle v1alpha1.PackageRevisionLifecycle) error {
-	ctx, span := tracer.Start(ctx, "gitRepository::UpdateLifecycle", trace.WithAttributes())
-	defer span.End()
-
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	old := r.getLifecycle(pkgRev)
-	if !v1alpha1.LifecycleIsPublished(old) {
-		return fmt.Errorf("cannot update lifecycle for draft package revision")
-	}
-	refSpecs := newPushRefSpecBuilder()
-	deletionProposedBranch := createDeletionProposedName(pkgRev.Key())
-
-	if old == v1alpha1.PackageRevisionLifecyclePublished {
-		if newLifecycle != v1alpha1.PackageRevisionLifecycleDeletionProposed {
-			return fmt.Errorf("invalid new lifecycle value: %q", newLifecycle)
-		}
-		// Push the package revision into a deletionProposed branch.
-		r.deletionProposedCache[deletionProposedBranch] = true
-		refSpecs.AddRefToPush(pkgRev.commit, deletionProposedBranch.RefInLocal())
-	}
-	if old == v1alpha1.PackageRevisionLifecycleDeletionProposed {
-		if newLifecycle != v1alpha1.PackageRevisionLifecyclePublished {
-			return fmt.Errorf("invalid new lifecycle value: %q", newLifecycle)
-		}
-
-		// Delete the deletionProposed branch
-		delete(r.deletionProposedCache, deletionProposedBranch)
-		ref := plumbing.NewHashReference(deletionProposedBranch.RefInLocal(), pkgRev.commit)
-		refSpecs.AddRefToDelete(ref)
-	}
-
-	if err := r.pushAndCleanup(ctx, refSpecs); err != nil {
-		if !pkgerrors.Is(err, git.NoErrAlreadyUpToDate) {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (r *gitRepository) UpdateDraftResources(ctx context.Context, draft *gitPackageRevisionDraft, new *v1alpha1.PackageRevisionResources, change *v1alpha1.Task) error {
-	ctx, span := tracer.Start(ctx, "gitRepository::UpdateResources", trace.WithAttributes())
-	defer span.End()
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	ch, err := newCommitHelper(r, r.userInfoProvider, draft.commit, draft.Key().PkgKey.ToFullPathname(), plumbing.ZeroHash)
-	if err != nil {
-		return pkgerrors.Wrap(err, "failed to commit package:")
-	}
-
-	for k, v := range new.Spec.Resources {
-		if err := ch.storeFile(filepath.Join(draft.Key().PkgKey.ToFullPathname(), k), v); err != nil {
-			return err
-		}
-	}
-
-	// Because we can't read the package back without a Kptfile, make sure one is present
-	{
-		p := filepath.Join(draft.Key().PkgKey.ToFullPathname(), "Kptfile")
-		_, err := ch.readFile(p)
-		if os.IsNotExist(err) {
-			// We could write the file here; currently we return an error
-			return pkgerrors.Wrap(err, "package must contain Kptfile at root")
-		}
-	}
-
-	annotation := &gitAnnotation{
-		PackagePath:   draft.Key().PkgKey.ToFullPathname(),
-		WorkspaceName: draft.Key().WorkspaceName,
-		Revision:      repository.Revision2Str(draft.Key().Revision),
-		Task:          change,
-	}
-	message := "Intermediate commit"
-	if change != nil {
-		message += fmt.Sprintf(": %s", change.Type)
-		draft.tasks = append(draft.tasks, *change)
-	}
-	message += "\n"
-
-	message, err = AnnotateCommitMessage(message, annotation)
-	if err != nil {
-		return err
-	}
-
-	commitHash, packageTree, err := ch.commit(ctx, message, draft.Key().PkgKey.ToFullPathname())
-	if err != nil {
-		return pkgerrors.Wrap(err, "failed to commit package: %w")
-	}
-
-	draft.tree = packageTree
-	draft.commit = commitHash
-	return nil
-}
-
 func (r *gitRepository) ClosePackageRevisionDraft(ctx context.Context, prd repository.PackageRevisionDraft, version int) (repository.PackageRevision, error) {
 	ctx, span := tracer.Start(ctx, "gitRepository::ClosePackageRevisionDraft", trace.WithAttributes())
 	defer span.End()
@@ -1600,28 +791,6 @@ func (r *gitRepository) ClosePackageRevisionDraft(ctx context.Context, prd repos
 	}, nil
 }
 
-// doGitWithAuth fetches auth information for git and provides it
-// to the provided function which performs the operation against a git repo.
-func (r *gitRepository) doGitWithAuth(ctx context.Context, op func(transport.AuthMethod) error) error {
-	auth, err := r.getAuthMethod(ctx, false)
-	if err != nil {
-		return fmt.Errorf("failed to obtain git credentials: %w", err)
-	}
-	err = op(auth)
-	if err != nil {
-		if !pkgerrors.Is(err, transport.ErrAuthenticationRequired) {
-			return err
-		}
-		klog.Infof("Authentication failed. Trying to refresh credentials")
-		// TODO: Consider having some kind of backoff here.
-		auth, err := r.getAuthMethod(ctx, true)
-		if err != nil {
-			return fmt.Errorf("failed to obtain git credentials: %w", err)
-		}
-		return op(auth)
-	}
-	return nil
-}
 
 func (r *gitRepository) commitPackageToMain(ctx context.Context, d *gitPackageRevisionDraft) (commitHash, newPackageTreeHash plumbing.Hash, base *plumbing.Reference, err error) {
 	ctx, span := tracer.Start(ctx, "gitRepository::commitPackageToMain", trace.WithAttributes())
@@ -1763,4 +932,46 @@ func getPkgWorkspace(commit *object.Commit, p *packageListEntry, ref *plumbing.R
 		}
 	}
 	return ""
+}
+
+// getAuthMethod fetches the credentials for authenticating to git. It caches the
+// credentials between calls and refresh credentials when the tokens have expired.
+func (r *gitRepository) getAuthMethod(ctx context.Context, forceRefresh bool) (transport.AuthMethod, error) {
+	// If no secret is provided, we try without any auth.
+	if r.secret == "" {
+		return nil, nil
+	}
+
+	if r.credential == nil || !r.credential.Valid() || forceRefresh {
+		if cred, err := r.credentialResolver.ResolveCredential(ctx, r.namespace, r.secret); err != nil {
+			return nil, fmt.Errorf("failed to obtain credential from secret %s/%s: %w", r.namespace, r.secret, err)
+		} else {
+			r.credential = cred
+		}
+	}
+
+	return r.credential.ToAuthMethod(), nil
+}
+
+// doGitWithAuth fetches auth information for git and provides it
+// to the provided function which performs the operation against a git repo.
+func (r *gitRepository) doGitWithAuth(ctx context.Context, op func(transport.AuthMethod) error) error {
+	auth, err := r.getAuthMethod(ctx, false)
+	if err != nil {
+		return fmt.Errorf("failed to obtain git credentials: %w", err)
+	}
+	err = op(auth)
+	if err != nil {
+		if !pkgerrors.Is(err, transport.ErrAuthenticationRequired) {
+			return err
+		}
+		klog.Infof("Authentication failed. Trying to refresh credentials")
+		// TODO: Consider having some kind of backoff here.
+		auth, err := r.getAuthMethod(ctx, true)
+		if err != nil {
+			return fmt.Errorf("failed to obtain git credentials: %w", err)
+		}
+		return op(auth)
+	}
+	return nil
 }
